usivc: libivc error handling and recovery
On UNIX socket IO, return appropriate error codes from errno.
This handles signals and transfer issue by attempting to re-open the
connection to ivcDaemon every 2 seconds automatically.

Switch to send to instruct the kernel not to send a SIGPIPE if
transmitting to a closed fd. In that optic, switch read to recv for a
more consistent low-level API.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>