module: do not change the cache policy on mmap
Xen PV domains are not allowed access to the PAT bits, this would
require a patch in Xen changing the l1_disallow_mask macro. It used to
be the case for xenfb2 to change the policy from a pv guest and allow
surfman to configure the GPU to scan these pages.

This is not what is happening in vGlass though, and the page cache
policy should not have an effect here.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>