usivc: packet-size operation and spurious ack
Remove the spurious ACK packet replied after receiving any control
packet from the frontend in processControlEvent. This hides any error
from the backend and tricks the frontend in assuming the backend is
conencted as the status replied is 0.

Introduce read_packet/write_packet primitive that read and write exactly
the amount of data give or fail with ENOSPC if the ring is empty/full.
This is done as such so the read or write locks are taken so at worst
data is only added (read) or only removed (write) from the ring, not
affecting the amount read/written.

Also remove double reads that would otherwise discard potentially
valuable data from the ring.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>