usivc: process as many ctrl event as possible
processControlEvent is called when the control event-channel is raised
in the frontend. All packets available have to be read since the
frontend will most likely not raise the event-channel again.

Note: when IVC services are initialized, the frontend sends a CONNECT
packet and expects an ACK before it times out and consider the attempt
to connect failed. Should the backend react "later" the resources shared
by the frontend are most likely no longer available.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>